excelium_app/resource_creator/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions==1.11.2
azure-storage-blob==12.12.0
azure-identity==1.10.0